+++ soxspipe/cl_utils.py
@@ -7,13 +7,16 @@ Usage:
     soxspipe init
     soxspipe mbias <inputFrames> [-s <pathToSettingsFile>] 
     soxspipe mdark <inputFrames> [-s <pathToSettingsFile>] 
+    soxspipe mflat <inputFrames> [-s <pathToSettingsFile>] 
     soxspipe disp_sol <inputFrames> [-s <pathToSettingsFile>] 
     soxspipe order_centres <inputFrames> [-s <pathToSettingsFile>] 
+
 
 Options:
     init                                   setup the soxspipe settings file for the first time
     mbias                                  the master bias recipe
     mdark                                  the master dark recipe
+    mflat                                  the master flat recipe
     disp_sol                               the disp solution recipe
     order_centres                          the order centres recipe
 
@@ -167,6 +170,16 @@ def main(arguments=None):
             inputFrames=a["inputFrames"]
         ).produce_product()
         print(f"\nThe order centre locations have been saved to an order table: {order_table}")
+
+    if a["mflat"]:
+        from soxspipe.recipes import soxs_mflat
+        recipe = soxs_mflat(
+            log=log,
+            settings=settings,
+            inputFrames=a["inputFrames"]
+        )
+        mflatFrame = recipe.produce_product()
+        print(f"\nThe master flat frame has been saved to: {mflatFrame}")
 
     # CALL FUNCTIONS/OBJECTS