+++ soxspipe/commonutils/__init__.py
@@ -7,5 +7,7 @@ from .detector_lookup import detector_lookup
 from .create_dispersion_map import create_dispersion_map
 from .getpackagepath import getpackagepath
 from .detect_continuum import detect_continuum
+from .detect_continuum import _base_detect
 from . import polynomials
 from . import toolkit
+from .detect_order_edges import detect_order_edges