+++ soxspipe/commonutils/toolkit.py
@@ -18,6 +18,8 @@ from soxspipe.commonutils.polynomials import chebyshev_xy_polynomial
 import unicodecsv as csv
 import numpy as np
 import matplotlib.pyplot as plt
+import random
+import numpy.ma as ma
 
 
 def cut_image_slice(
@@ -27,44 +29,57 @@ def cut_image_slice(
         length,
         x,
         y,
-        plot):
-    """*cut and return an N-pixel wide and M-pixels long slice, centred on a given coordinate from an image frame*
+        median=False,
+        plot=False):
+    """*cut and return an N-pixel wide and M-pixels long slice, centred on a given coordinate from an image frame. Also return the origin pixel value for the slice*
 
     **Key Arguments:**
 
     - ``log`` -- logger
-    - ``frame`` -- the data array to cut the slice from 
-    - ``width`` -- width of the slice
+    - ``frame`` -- the data array to cut the slice from (masked array)
+    - ``width`` -- width of the slice (odd number)
     - ``length`` -- length of the slice
     - ``x`` -- x-coordinate
     - ``y`` -- y-coordinate
+    - ``median`` -- collapse the slice to a median value across its width
     - ``plot`` -- generate a plot of slice. Useful for debugging.
 
     **Usage:**
 
-    ```eval_rst
-    .. todo::
-
-            add usage info
-            create a sublime snippet for usage
-    ```
-
     ```python
-    usage code 
+    from soxspipe.commonutils.toolkit import cut_image_slice
+    slice = cut_image_slice(log=self.log, frame=self.pinholeFlat.data,
+                                    width=1, length=sliceLength, x=x_fit, y=y_fit, plot=False)
+    if slice is None:
+        return None
     ```           
     """
     log.debug('starting the ``cut_image_slice`` function')
 
-    print(frame.shape)
+    # print(frame.shape)
     halfSlice = length / 2
 
+    #
+    if (width % 2) != 0:
+        halfwidth = (width - 1) / 2
+    else:
+        halfwidth = width / 2
+
     # CHECK WE ARE NOT GOING BEYOND BOUNDS OF FRAME
-    if (x > frame.shape[1] - halfSlice) or (y > frame.shape[0]):
+    if (x > frame.shape[1] - halfSlice) or (y > frame.shape[0] - halfwidth) or (x < halfSlice) or (y < halfwidth):
         return None
 
-    slice = frame[int(y), int(x - halfSlice):int(x + halfSlice)]
+    slice = frame[int(y - halfwidth):int(y + halfwidth + 1),
+                  int(x - halfSlice):int(x + halfSlice)]
 
-    if plot:
+    # print(int(x - halfSlice), y)
+    # print(slice.data.shape)
+
+    if median:
+        slice = ma.median(slice, axis=0)
+        # print(slice.mask)
+
+    if plot and random.randint(1, 101) < 10000:
         # CHECK THE SLICE POINTS IF NEEDED
         x = np.arange(0, len(slice))
         plt.figure(figsize=(8, 5))
@@ -105,6 +120,8 @@ def quicklook_image(
 
     frame = CCDObject
     rotatedImg = np.rot90(frame, 1)
+    rotatedImg = np.flipud(np.rot90(frame, 1))
+
     std = np.std(frame.data)
     mean = np.mean(frame.data)
     vmax = mean + 3 * std
@@ -117,6 +134,7 @@ def quicklook_image(
         "y-axis", fontsize=10)
     plt.ylabel(
         "x-axis", fontsize=10)
+    plt.invert_yaxis()
     plt.show()
 
     log.debug('completed the ``quicklook_image`` function')
@@ -140,7 +158,7 @@ def unpack_order_table(
     ```python
     # UNPACK THE ORDER TABLE
     from soxspipe.commonutils.toolkit import unpack_order_table
-    orderCentres = unpack_order_table(
+    orders, orderCentres, orderLimits = unpack_order_table(
         log=self.log, orderTablePath=orderTablePath)
     ```           
     """
@@ -148,23 +166,27 @@ def unpack_order_table(
 
     # OPEN ORDER TABLE AND GENERATE PIXEL ARRAYS FOR CENTRE LOCATIONS
     orderCentres = []
+    orders = []
+    orderLimits = []
 
     with open(orderTablePath, 'rb') as csvFile:
         csvReader = csv.DictReader(
             csvFile, dialect='excel', delimiter=',', quotechar='"')
         for row in csvReader:
             order = int(row["order"])
-            degy = int(row["degy"])
+            degy_cent = int(row["degy_cent"])
             ymin = int(row["ymin"])
             ymax = int(row["ymax"])
             cent_coeff = [float(v) for k, v in row.items() if "CENT_" in k]
             ycoords = np.arange(ymin, ymax, 1)
             poly = chebyshev_xy_polynomial(
-                log=log, deg=degy).poly
+                log=log, deg=degy_cent).poly
             xcoords = poly(ycoords, *cent_coeff)
             orderCentres.append((xcoords, ycoords))
+            orders.append(order)
+            orderLimits.append((ymin, ymax))
 
         csvFile.close()
 
     log.debug('completed the ``functionName`` function')
-    return orderCentres
+    return orders, orderCentres, orderLimits